move cocos/core/platform/view.ts to cocos/ui/view.ts
