+++ cocos/2d/framework/canvas.ts
@@ -30,7 +30,7 @@ import { EDITOR } from 'internal:constants';
 import { Camera } from '../../misc/camera-component';
 import { Widget } from '../../ui/widget';
 import { Vec3 } from '../../core/math';
-import { view } from '../../core/platform/view';
+import { view } from '../../ui/view';
 import { legacyCC } from '../../core/global-exports';
 import { Enum } from '../../core/value-types/enum';
 import visibleRect from '../../core/platform/visible-rect';
