+++ cocos/ui/deprecated.ts
@@ -43,6 +43,8 @@ import { warnID } from '../core/platform/debug';
 import { ccclass } from '../core/data/class-decorator';
 import { js } from '../core/utils/js';
 import { legacyCC } from '../core/global-exports';
+import { View } from './view';
+import { removeProperty, markAsWarning } from '../core';
 
 /**
  * @deprecated Since v1.2
@@ -159,3 +161,59 @@ js.setClassAlias(UICoordinateTracker, 'cc.UICoordinateTrackerComponent');
 export { BlockInputEvents as BlockInputEventsComponent };
 legacyCC.BlockInputEventsComponent = BlockInputEvents;
 js.setClassAlias(BlockInputEvents, 'cc.BlockInputEventsComponent');
+
+// #region deprecation on view
+removeProperty(View.prototype, 'View.prototype', [
+    {
+        name: 'isAntiAliasEnabled',
+        suggest: 'The API of Texture2d have been largely modified, no alternative',
+    },
+    {
+        name: 'enableAntiAlias',
+        suggest: 'The API of Texture2d have been largely modified, no alternative',
+    },
+]);
+markAsWarning(View.prototype, 'View.prototype', [
+    {
+        name: 'adjustViewportMeta',
+    },
+    {
+        name: 'enableAutoFullScreen',
+        suggest: 'use screen.requestFullScreen() instead.',
+    },
+    {
+        name: 'isAutoFullScreenEnabled',
+    },
+    {
+        name: 'setCanvasSize',
+        suggest: 'setting size in CSS pixels is not recommended, please use screen.windowSize instead.',
+    },
+    {
+        name: 'getCanvasSize',
+        suggest: 'please use screen.windowSize instead.',
+    },
+    {
+        name: 'getFrameSize',
+        suggest: 'getting size in CSS pixels is not recommended, please use screen.windowSize instead.',
+    },
+    {
+        name: 'setFrameSize',
+        suggest: 'setting size in CSS pixels is not recommended, please use screen.windowSize instead.',
+    },
+    {
+        name: 'getDevicePixelRatio',
+        suggest: 'use screen.devicePixelRatio instead.',
+    },
+    {
+        name: 'convertToLocationInView',
+    },
+    {
+        name: 'enableRetina',
+    },
+    {
+        name: 'isRetinaEnabled',
+    },
+    {
+        name: 'setRealPixelResolution',
+    },
+]);
