+++ cocos/ui/index.ts
@@ -41,4 +41,5 @@ export { UICoordinateTracker } from './ui-coordinate-tracker';
 export { BlockInputEvents } from './block-input-events';
 export { SubContextView } from './sub-context-view';
 export { widgetManager } from './widget-manager';
+export * from './view';
 export * from './deprecated';
