+++ cocos/ui/safe-area.ts
@@ -28,10 +28,11 @@ import { EDITOR } from 'internal:constants';
 import { screenAdapter } from 'pal/screen-adapter';
 import { Component } from '../scene-graph/component';
 import { UITransform } from '../2d/framework';
-import { view, sys } from '../core/platform';
+import { sys } from '../core/platform';
 import { Widget } from './widget';
 import { widgetManager } from './widget-manager';
 import { legacyCC } from '../core/global-exports';
+import { view } from './view';
 
 /**
  * @en
