+++ cocos/ui/widget-manager.ts
@@ -28,7 +28,7 @@ import { EDITOR, DEV } from 'internal:constants';
 import { screenAdapter } from 'pal/screen-adapter';
 import { Director, director } from '../game/director';
 import { Vec2, Vec3 } from '../core/math';
-import { View } from '../core/platform/view';
+import { View } from './view';
 import visibleRect from '../core/platform/visible-rect';
 import { Scene } from '../scene-graph';
 import { Node } from '../scene-graph/node';
