+++ tests/core/view.test.ts
@@ -1,6 +1,6 @@
 import { screenAdapter } from 'pal/screen-adapter';
 import { Rect, screen, Size, sys, Vec2 } from '../../cocos/core';
-import { view, ResolutionPolicy } from '../../cocos/core/platform/view';
+import { view, ResolutionPolicy } from '../../cocos/ui/view';
 
 describe('cc.view', () => {
     // init environment
